crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
supabase==2.4.2
python-dotenv==1.0.1
playwright==1.43.0
pyyaml==6.0.1
